% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piledownR.R
\name{generate}
\alias{generate}
\title{Generate results from a PileParams or JunctionParams object.}
\usage{
generate(params)
}
\arguments{
\item{params}{A PileParams or JunctionParams object.}
}
\value{
An \code{arrow::RecordBatchReader}.
}
\description{
For PileParams, returns an Arrow RecordBatchReader with columns:
name, seq, strand, pos, up, down.
For JunctionParams, returns an Arrow RecordBatchReader with columns:
name, seq, strand, start, end, count.
}
